+++ tests.py
@@ -1,41 +1,73 @@
 from __future__ import unicode_literals
 from decimal import Decimal
-from io import BytesIO
+import contextlib
 import glob
+import io
+import json
 import os
 import re
+import subprocess
 import sys
-import json
 
 if sys.version_info[:2] < (2, 7):  # pragma: no cover
-    from unittest2 import TestCase, skipIf
+    import unittest2 as unittest
 else:
-    from unittest import TestCase, skipIf
+    import unittest
 
 try:
     from unittest import mock
 except ImportError:
     import mock
 
+try:
+    from sys import pypy_version_info
+except ImportError:
+    pypy_version_info = None
+
 from jsonschema import (
     PY3, SchemaError, UnknownType, ValidationError, ErrorTree,
-    Draft3Validator, RefResolver, validate
+    Draft3Validator, Draft4Validator, FormatChecker, draft3_format_checker,
+    draft4_format_checker, RefResolver, validate, FormatError
 )
 
 
-def make_case(schema, data, valid, cls):
-    def test_case(self):
-        if valid:
-            validate(data, schema, cls=cls)
-        else:
+THIS_DIR = os.path.dirname(__file__)
+TESTS_DIR = os.path.join(THIS_DIR, "json", "tests")
+
+JSONSCHEMA_SUITE = os.path.join(THIS_DIR, "json", "bin", "jsonschema_suite")
+
+REMOTES = subprocess.Popen(
+    ["python", JSONSCHEMA_SUITE, "remotes"], stdout=subprocess.PIPE,
+).stdout
+if PY3:
+    REMOTES = io.TextIOWrapper(REMOTES)
+REMOTES = json.load(REMOTES)
+
+
+def make_case(schema, data, valid):
+    if valid:
+        def test_case(self):
+            kwargs = getattr(self, "validator_kwargs", {})
+            validate(data, schema, cls=self.validator_class, **kwargs)
+    else:
+        def test_case(self):
+            kwargs = getattr(self, "validator_kwargs", {})
             with self.assertRaises(ValidationError):
-                validate(data, schema, cls=cls)
+                validate(data, schema, cls=self.validator_class, **kwargs)
     return test_case
 
 
-def load_json_cases(test_dir):
+def load_json_cases(tests_glob, ignore_glob="", basedir=TESTS_DIR, skip=None):
+    if ignore_glob:
+        ignore_glob = os.path.join(basedir, ignore_glob)
+
     def add_test_methods(test_class):
-        for filename in glob.iglob(os.path.join(test_dir, "*.json")):
+        ignored = set(glob.iglob(ignore_glob))
+
+        for filename in glob.iglob(os.path.join(basedir, tests_glob)):
+            if filename in ignored:
+                continue
+
             validating, _ = os.path.splitext(os.path.basename(filename))
 
             with open(filename) as test_file:
@@ -47,7 +79,6 @@ def load_json_cases(test_dir):
                             case["schema"],
                             test["data"],
                             test["valid"],
-                            test_class.validator_class,
                         )
 
                         test_name = "test_%s_%s" % (
@@ -59,14 +90,19 @@ def load_json_cases(test_dir):
                             test_name = test_name.encode("utf-8")
                         a_test.__name__ = test_name
 
+                        if skip is not None and skip(case):
+                            a_test = unittest.skip("Checker not present.")(
+                                a_test
+                            )
+
                         setattr(test_class, test_name, a_test)
 
         return test_class
     return add_test_methods
 
 
-class ByteStringMixin(object):
-    @skipIf(PY3, "The JSON module in Python 3 always produces unicode")
+class TypesMixin(object):
+    @unittest.skipIf(PY3, "In Python 3 json.load always produces unicode")
     def test_string_a_bytestring_is_a_string(self):
         self.validator_class({"type" : "string"}).validate(b"foo")
 
@@ -86,27 +122,128 @@ class DecimalMixin(object):
                 validator.validate(invalid)
 
 
-class AnyTypeMixin(object):
+def missing_format(checker):
+    def missing_format(case):
+        format = case["schema"].get("format")
+        return format not in checker.checkers or (
+            # datetime.datetime is overzealous about typechecking in <=1.9
+            format == "date-time" and
+            pypy_version_info is not None and
+            pypy_version_info[:2] <= (1, 9)
+        )
+    return missing_format
+
+
+class FormatMixin(object):
+    def test_it_returns_true_for_formats_it_does_not_know_about(self):
+        validator = self.validator_class(
+            {"format" : "carrot"}, format_checker=FormatChecker(),
+        )
+        validator.validate("bugs")
+
+    def test_it_does_not_validate_formats_by_default(self):
+        validator = self.validator_class({})
+        self.assertIsNone(validator.format_checker)
+
+    def test_it_validates_formats_if_a_checker_is_provided(self):
+        checker = mock.Mock(spec=FormatChecker)
+        validator = self.validator_class(
+            {"format" : "foo"}, format_checker=checker,
+        )
+
+        validator.validate("bar")
+
+        checker.check.assert_called_once_with("bar", "foo")
+
+        cause = ValueError()
+        checker.check.side_effect = FormatError('aoeu', cause=cause)
+
+        with self.assertRaises(ValidationError) as cm:
+            validator.validate("bar")
+        # Make sure original cause is attached
+        self.assertIs(cm.exception.cause, cause)
+
+
+@load_json_cases(
+    "draft3/*.json", ignore_glob=os.path.join("draft3", "refRemote.json")
+)
+@load_json_cases(
+    "draft3/optional/format.json", skip=missing_format(draft3_format_checker)
+)
+@load_json_cases("draft3/optional/bignum.json")
+@load_json_cases("draft3/optional/zeroTerminatedFloats.json")
+class TestDraft3(
+    unittest.TestCase, TypesMixin, DecimalMixin, FormatMixin
+):
+
+    validator_class = Draft3Validator
+    validator_kwargs = {"format_checker" : draft3_format_checker}
+
     def test_any_type_is_valid_for_type_any(self):
         validator = self.validator_class({"type" : "any"})
         validator.validate(mock.Mock())
 
+    # TODO: we're in need of more meta schema tests
+    def test_invalid_properties(self):
+        with self.assertRaises(SchemaError):
+            validate({}, {"properties": {"test": True}},
+                     cls=self.validator_class)
 
-@load_json_cases(os.path.join(os.path.dirname(__file__), "json/tests/draft3/"))
-class TestDraft3(TestCase, ByteStringMixin, DecimalMixin, AnyTypeMixin):
-    validator_class = Draft3Validator
+    def test_minItems_invalid_string(self):
+        with self.assertRaises(SchemaError):
+            # needs to be an integer
+            validate([1], {"minItems" : "1"}, cls=self.validator_class)
+
+
+@load_json_cases(
+    "draft4/*.json", ignore_glob=os.path.join("draft4", "refRemote.json")
+)
+@load_json_cases(
+    "draft4/optional/format.json", skip=missing_format(draft4_format_checker)
+)
+@load_json_cases("draft4/optional/bignum.json")
+@load_json_cases("draft4/optional/zeroTerminatedFloats.json")
+class TestDraft4(
+    unittest.TestCase, TypesMixin, DecimalMixin, FormatMixin
+):
+    validator_class = Draft4Validator
+    validator_kwargs = {"format_checker" : draft4_format_checker}
 
     # TODO: we're in need of more meta schema tests
     def test_invalid_properties(self):
         with self.assertRaises(SchemaError):
-            validate({}, {"properties": {"test": True}})
+            validate({}, {"properties": {"test": True}},
+                     cls=self.validator_class)
 
     def test_minItems_invalid_string(self):
         with self.assertRaises(SchemaError):
-            validate([1], {"minItems" : "1"})  # needs to be an integer
+            # needs to be an integer
+            validate([1], {"minItems" : "1"}, cls=self.validator_class)
 
 
-class TestIterErrors(TestCase):
+class RemoteRefResolution(unittest.TestCase):
+    def setUp(self):
+        patch = mock.patch("jsonschema.requests")
+        requests = patch.start()
+        requests.get.side_effect = self.resolve
+        self.addCleanup(patch.stop)
+
+    def resolve(self, reference):
+        _, _, reference = reference.partition("http://localhost:1234/")
+        return mock.Mock(**{"json.return_value" : REMOTES.get(reference)})
+
+
+@load_json_cases("draft3/refRemote.json")
+class Draft3RemoteResolution(RemoteRefResolution):
+    validator_class = Draft3Validator
+
+
+@load_json_cases("draft4/refRemote.json")
+class Draft4RemoteResolution(RemoteRefResolution):
+    validator_class = Draft4Validator
+
+
+class TestIterErrors(unittest.TestCase):
     def setUp(self):
         self.validator = Draft3Validator({})
 
@@ -140,10 +277,11 @@ class TestIterErrors(TestCase):
         self.assertEqual(len(errors), 4)
 
 
-class TestValidationErrorMessages(TestCase):
-    def message_for(self, instance, schema):
+class TestValidationErrorMessages(unittest.TestCase):
+    def message_for(self, instance, schema, *args, **kwargs):
+        kwargs.setdefault("cls", Draft3Validator)
         with self.assertRaises(ValidationError) as e:
-            validate(instance, schema)
+            validate(instance, schema, *args, **kwargs)
         return e.exception.message
 
     def test_single_type_failure(self):
@@ -202,8 +340,20 @@ class TestValidationErrorMessages(TestCase):
         self.assertIn(repr("bar"), message)
         self.assertIn("were unexpected)", message)
 
+    def test_invalid_format_default_message(self):
+        checker = FormatChecker(formats=())
+        check_fn = mock.Mock(return_value=False)
+        checker.checks("thing")(check_fn)
 
-class TestValidationErrorDetails(TestCase):
+        schema = {"format" : "thing"}
+        message = self.message_for("bla", schema, format_checker=checker)
+
+        self.assertIn(repr("bla"), message)
+        self.assertIn(repr("thing"), message)
+        self.assertIn("is not a", message)
+
+
+class TestValidationErrorDetails(unittest.TestCase):
     def setUp(self):
         self.validator = Draft3Validator({})
 
@@ -222,10 +372,10 @@ class TestValidationErrorDetails(TestCase):
         errors = self.validator.iter_errors(instance, schema)
         e1, e2, e3, e4 = sorted_errors(errors)
 
-        self.assertEqual(e1.path, ["bar"])
-        self.assertEqual(e2.path, ["baz"])
-        self.assertEqual(e3.path, ["baz"])
-        self.assertEqual(e4.path, ["foo"])
+        self.assertEqual(list(e1.path), ["bar"])
+        self.assertEqual(list(e2.path), ["baz"])
+        self.assertEqual(list(e3.path), ["baz"])
+        self.assertEqual(list(e4.path), ["foo"])
 
         self.assertEqual(e1.validator, "minItems")
         self.assertEqual(e2.validator, "enum")
@@ -254,12 +404,12 @@ class TestValidationErrorDetails(TestCase):
         errors = self.validator.iter_errors(instance, schema)
         e1, e2, e3, e4, e5, e6 = sorted_errors(errors)
 
-        self.assertEqual(e1.path, [])
-        self.assertEqual(e2.path, [0])
-        self.assertEqual(e3.path, ["bar", 1])
-        self.assertEqual(e4.path, ["bar", "bar", 1])
-        self.assertEqual(e5.path, ["baz", "bar", 1])
-        self.assertEqual(e6.path, ["foo", 1])
+        self.assertEqual(list(e1.path), [])
+        self.assertEqual(list(e2.path), [0])
+        self.assertEqual(list(e3.path), [1, "bar"])
+        self.assertEqual(list(e4.path), [1, "bar", "bar"])
+        self.assertEqual(list(e5.path), [1, "bar", "baz"])
+        self.assertEqual(list(e6.path), [1, "foo"])
 
         self.assertEqual(e1.validator, "type")
         self.assertEqual(e2.validator, "type")
@@ -268,8 +418,57 @@ class TestValidationErrorDetails(TestCase):
         self.assertEqual(e5.validator, "minItems")
         self.assertEqual(e6.validator, "enum")
 
+    def test_additionalProperties(self):
+        instance = {"bar": "bar", "foo": 2}
+        schema = {
+            "additionalProperties" : {"type": "integer", "minimum": 5}
+        }
 
-class TestErrorTree(TestCase):
+        errors = self.validator.iter_errors(instance, schema)
+        e1, e2 = sorted_errors(errors)
+
+        self.assertEqual(list(e1.path), ["bar"])
+        self.assertEqual(list(e2.path), ["foo"])
+
+        self.assertEqual(e1.validator, "type")
+        self.assertEqual(e2.validator, "minimum")
+
+    def test_patternProperties(self):
+        instance = {"bar": 1, "foo": 2}
+        schema = {
+            "patternProperties" : {
+                "bar": {"type": "string"},
+                "foo": {"minimum": 5}
+            }
+        }
+
+        errors = self.validator.iter_errors(instance, schema)
+        e1, e2 = sorted_errors(errors)
+
+        self.assertEqual(list(e1.path), ["bar"])
+        self.assertEqual(list(e2.path), ["foo"])
+
+        self.assertEqual(e1.validator, "type")
+        self.assertEqual(e2.validator, "minimum")
+
+    def test_additionalItems(self):
+        instance = ["foo", 1]
+        schema = {
+            "items": [],
+            "additionalItems" : {"type": "integer", "minimum": 5}
+        }
+
+        errors = self.validator.iter_errors(instance, schema)
+        e1, e2 = sorted_errors(errors)
+
+        self.assertEqual(list(e1.path), [0])
+        self.assertEqual(list(e2.path), [1])
+
+        self.assertEqual(e1.validator, "type")
+        self.assertEqual(e2.validator, "minimum")
+
+
+class TestErrorTree(unittest.TestCase):
     def setUp(self):
         self.validator = Draft3Validator({})
 
@@ -296,7 +495,7 @@ class TestErrorTree(TestCase):
     def test_it_creates_a_child_tree_for_each_nested_path(self):
         errors = [
             ValidationError("a bar message", path=["bar"]),
-            ValidationError("a bar -> 0 message", path=[0, "bar"]),
+            ValidationError("a bar -> 0 message", path=["bar", 0]),
         ]
         tree = ErrorTree(errors)
         self.assertIn(0, tree["bar"])
@@ -304,19 +503,19 @@ class TestErrorTree(TestCase):
 
     def test_children_have_their_errors_dicts_built(self):
         e1, e2 = (
-            ValidationError("message 1", validator="foo", path=[0, "bar"]),
-            ValidationError("message 2", validator="quux", path=[0, "bar"]),
+            ValidationError("message 1", validator="foo", path=["bar", 0]),
+            ValidationError("message 2", validator="quux", path=["bar", 0]),
         )
         tree = ErrorTree([e1, e2])
         self.assertEqual(tree["bar"][0].errors, {"foo" : e1, "quux" : e2})
 
 
-class TestDraft3Validator(TestCase):
+class ValidatorTestMixin(object):
     def setUp(self):
         self.instance = mock.Mock()
         self.schema = {}
         self.resolver = mock.Mock()
-        self.validator = Draft3Validator(self.schema)
+        self.validator = self.validator_class(self.schema)
 
     def test_valid_instances_are_valid(self):
         errors = iter([])
@@ -346,14 +545,19 @@ class TestDraft3Validator(TestCase):
         self.assertIsInstance(self.validator.resolver, RefResolver)
 
     def test_it_delegates_to_a_ref_resolver(self):
-        resolver = mock.Mock()
-        resolver.resolve.return_value = {"type" : "integer"}
+        resolver = RefResolver("", {})
         schema = {"$ref" : mock.Mock()}
 
-        with self.assertRaises(ValidationError):
-            Draft3Validator(schema, resolver=resolver).validate(None)
+        @contextlib.contextmanager
+        def resolving():
+            yield {"type": "integer"}
 
-        resolver.resolve.assert_called_once_with(schema["$ref"])
+        with mock.patch.object(resolver, "resolving") as resolve:
+            resolve.return_value = resolving()
+            with self.assertRaises(ValidationError):
+                self.validator_class(schema, resolver=resolver).validate(None)
+
+        resolve.assert_called_once_with(schema["$ref"])
 
     def test_is_type_is_true_for_valid_type(self):
         self.assertTrue(self.validator.is_type("foo", "string"))
@@ -361,23 +565,55 @@ class TestDraft3Validator(TestCase):
     def test_is_type_is_false_for_invalid_type(self):
         self.assertFalse(self.validator.is_type("foo", "array"))
 
-    def test_is_type_is_true_for_any_type(self):
-        self.assertTrue(self.validator.is_type(mock.Mock(), "any"))
-
     def test_is_type_evades_bool_inheriting_from_int(self):
         self.assertFalse(self.validator.is_type(True, "integer"))
         self.assertFalse(self.validator.is_type(True, "number"))
-
-    def test_is_type_does_not_evade_bool_if_it_is_being_tested(self):
-        self.assertTrue(self.validator.is_type(True, "boolean"))
-        self.assertTrue(self.validator.is_type(True, "any"))
 
     def test_is_type_raises_exception_for_unknown_type(self):
         with self.assertRaises(UnknownType):
             self.validator.is_type("foo", object())
 
 
-class TestRefResolver(TestCase):
+class TestDraft3Validator(ValidatorTestMixin, unittest.TestCase):
+    validator_class = Draft3Validator
+
+    def test_is_type_is_true_for_any_type(self):
+        self.assertTrue(self.validator.is_valid(mock.Mock(), {"type": "any"}))
+
+    def test_is_type_does_not_evade_bool_if_it_is_being_tested(self):
+        self.assertTrue(self.validator.is_type(True, "boolean"))
+        self.assertTrue(self.validator.is_valid(True, {"type": "any"}))
+
+
+class TestDraft4Validator(ValidatorTestMixin, unittest.TestCase):
+    validator_class = Draft4Validator
+
+
+class TestValidate(unittest.TestCase):
+    def test_draft3_validator_is_chosen(self):
+        schema = {"$schema" : "http://json-schema.org/draft-03/schema#"}
+        with mock.patch.object(Draft3Validator, "check_schema") as chk_schema:
+            validate({}, schema)
+            chk_schema.assert_called_once_with(schema)
+        # Make sure it works without the empty fragment
+        schema = {"$schema" : "http://json-schema.org/draft-03/schema"}
+        with mock.patch.object(Draft3Validator, "check_schema") as chk_schema:
+            validate({}, schema)
+            chk_schema.assert_called_once_with(schema)
+
+    def test_draft4_validator_is_chosen(self):
+        schema = {"$schema" : "http://json-schema.org/draft-04/schema#"}
+        with mock.patch.object(Draft4Validator, "check_schema") as chk_schema:
+            validate({}, schema)
+            chk_schema.assert_called_once_with(schema)
+
+    def test_draft4_validator_is_the_default(self):
+        with mock.patch.object(Draft4Validator, "check_schema") as chk_schema:
+            validate({}, {})
+            chk_schema.assert_called_once_with({})
+
+
+class TestRefResolver(unittest.TestCase):
     def setUp(self):
         self.base_uri = ""
         self.referrer = {}
@@ -387,31 +623,41 @@ class TestRefResolver(TestCase):
     def test_it_does_not_retrieve_schema_urls_from_the_network(self):
         ref = Draft3Validator.META_SCHEMA["id"]
         with mock.patch.object(self.resolver, "resolve_remote") as remote:
-            resolved = self.resolver.resolve(ref)
-
-        self.assertEqual(resolved, Draft3Validator.META_SCHEMA)
+            with self.resolver.resolving(ref) as resolved:
+                self.assertEqual(resolved, Draft3Validator.META_SCHEMA)
         self.assertFalse(remote.called)
 
     def test_it_resolves_local_refs(self):
         ref = "#/properties/foo"
         self.referrer["properties"] = {"foo" : object()}
-        resolved = self.resolver.resolve(ref)
-        self.assertEqual(resolved, self.referrer["properties"]["foo"])
+        with self.resolver.resolving(ref) as resolved:
+            self.assertEqual(resolved, self.referrer["properties"]["foo"])
 
     def test_it_retrieves_stored_refs(self):
-        ref = self.resolver.store["cached_ref"] = {"foo" : 12}
-        resolved = self.resolver.resolve("cached_ref#/foo")
-        self.assertEqual(resolved, 12)
+        self.resolver.store["cached_ref"] = {"foo" : 12}
+        with self.resolver.resolving("cached_ref#/foo") as resolved:
+            self.assertEqual(resolved, 12)
+
+    def test_it_retrieves_unstored_refs_via_requests(self):
+        ref = "http://bar#baz"
+        schema = {"baz" : 12}
+
+        with mock.patch("jsonschema.requests") as requests:
+            requests.get.return_value.json.return_value = schema
+            with self.resolver.resolving(ref) as resolved:
+                self.assertEqual(resolved, 12)
+        requests.get.assert_called_once_with("http://bar")
 
     def test_it_retrieves_unstored_refs_via_urlopen(self):
         ref = "http://bar#baz"
         schema = {"baz" : 12}
 
-        with mock.patch("jsonschema.urlopen") as urlopen:
-            urlopen.return_value.read.return_value = json.dumps(schema)
-            resolved = self.resolver.resolve(ref)
-
-        self.assertEqual(resolved, 12)
+        with mock.patch("jsonschema.requests", None):
+            with mock.patch("jsonschema.urlopen") as urlopen:
+                urlopen.return_value.read.return_value = (
+                    json.dumps(schema).encode("utf8"))
+                with self.resolver.resolving(ref) as resolved:
+                    self.assertEqual(resolved, 12)
         urlopen.assert_called_once_with("http://bar")
 
     def test_it_can_construct_a_base_uri_from_a_schema(self):
@@ -426,7 +672,81 @@ class TestRefResolver(TestCase):
         self.assertEqual(resolver.base_uri, "")
         self.assertEqual(resolver.referrer, schema)
 
+    def test_custom_uri_scheme_handlers(self):
+        schema = {"foo": "bar"}
+        ref = "foo://bar"
+        foo_handler = mock.Mock(return_value=schema)
+        resolver = RefResolver("", {}, handlers={"foo": foo_handler})
+        with resolver.resolving(ref) as resolved:
+            self.assertEqual(resolved, schema)
+        foo_handler.assert_called_once_with(ref)
+
+    def test_cache_remote_on(self):
+        ref = "foo://bar"
+        foo_handler = mock.Mock()
+        resolver = RefResolver("", {}, cache_remote=True,
+                               handlers={"foo": foo_handler})
+        with resolver.resolving(ref):
+            pass
+        with resolver.resolving(ref):
+            pass
+        foo_handler.assert_called_once_with(ref)
+
+    def test_cache_remote_off(self):
+        ref = "foo://bar"
+        foo_handler = mock.Mock()
+        resolver = RefResolver("", {}, cache_remote=False,
+                               handlers={"foo": foo_handler})
+        with resolver.resolving(ref):
+            pass
+        with resolver.resolving(ref):
+            pass
+        self.assertEqual(foo_handler.call_count, 2)
+
+
+class TestFormatChecker(unittest.TestCase):
+    def setUp(self):
+        self.fn = mock.Mock()
+
+    def test_it_can_validate_no_formats(self):
+        checker = FormatChecker(formats=())
+        self.assertFalse(checker.checkers)
+
+    def test_it_raises_a_key_error_for_unknown_formats(self):
+        with self.assertRaises(KeyError):
+            FormatChecker(formats=["o noes"])
+
+    def test_it_can_register_cls_checkers(self):
+        with mock.patch.dict(FormatChecker.checkers, clear=True):
+            FormatChecker.cls_checks("new")(self.fn)
+            self.assertEqual(FormatChecker.checkers, {"new" : (self.fn, ())})
+
+    def test_it_can_register_checkers(self):
+        checker = FormatChecker()
+        checker.checks("new")(self.fn)
+        self.assertEqual(
+            checker.checkers,
+            dict(FormatChecker.checkers, new=(self.fn, ()))
+        )
+
+    def test_it_catches_registered_errors(self):
+        checker = FormatChecker()
+        checker.checks("foo", raises=ValueError)(self.fn)
+        # Registered errors should be caught and turned into FormatErrors
+        cause = ValueError()
+        self.fn.side_effect = cause
+        with self.assertRaises(FormatError) as cm:
+            checker.check("bar", "foo")
+        # Original exception should be attached to cause attribute
+        self.assertIs(cm.exception.cause, cause)
+        # Unregistered errors should not be caught
+        self.fn.side_effect = AttributeError
+        with self.assertRaises(AttributeError):
+            checker.check("bar", "foo")
+
+
 
 def sorted_errors(errors):
-    def key(error) : return ([str(e) for e in error.path], error.validator)
+    def key(error):
+        return [str(e) for e in error.path], error.validator
     return sorted(errors, key=key)